+++ src/socket_io.h
@@ -52,7 +52,7 @@
  *     - R:         number of successfully transmitted bytes
  * 
  *******************************************************************************/
-ssize_t http_send_all( int socket, const void* buffer, size_t length, int flags ) ;
+long http_send_all( int socket, const void* buffer, long length, int flags );
 
 
 
@@ -74,7 +74,7 @@ ssize_t http_send_all( int socket, const void* buffer, size_t length, int flags 
  *                  -1 in case of other error
  * 
  *******************************************************************************/
-ssize_t http_recv_timedout( int socket, void* buffer, size_t length, 
+long http_recv_timedout( int socket, void* buffer, long length, 
   int flags, int timeout);
